Test for a single scoped namespace
